metadata: Fix metadata association problems concerning children processes (#1232)
* Resolve all the adjacent processes IDs in the same PID namespace


* refactor: Move funtions to semantically correct package


* Clean up


* [pre-commit.ci lite] apply automatic fixes

* Upgrade licence text


* [pre-commit.ci lite] apply automatic fixes

* Use 2022


* Remove TODO


* Remove TODOs


* Remove unnecessary changes


* [pre-commit.ci lite] apply automatic fixes

* Remove TODOs
